[#02][be][user]refactor: 구조 변경에 의한 수정
