+++ back/src/main/java/com/rouby/user/application/service/UserWriteService.java
@@ -1,12 +1,16 @@
 package com.rouby.user.application.service;
 
+import static com.rouby.user.application.exception.UserErrorCode.INVALID_EMAIL_VERIFICATION;
+
 import com.rouby.user.application.dto.SaveVerificationCodeCommand;
+import com.rouby.user.application.dto.VerifyEmailCommand;
 import com.rouby.user.application.dto.command.CreateUserCommand;
 import com.rouby.user.application.exception.UserErrorCode;
 import com.rouby.user.application.exception.UserException;
 import com.rouby.user.domain.entity.User;
+import com.rouby.user.application.service.verification.VerificationEmailCode;
 import com.rouby.user.domain.repository.UserRepository;
-import com.rouby.user.domain.repository.VerificationEmailCodeRepository;
+import com.rouby.user.application.service.verification.VerificationEmailCodeStorage;
 import com.rouby.user.domain.service.UserPasswordEncoder;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -18,7 +22,7 @@ public class UserWriteService {
 
   private final UserRepository userRepository;
   private final UserPasswordEncoder passwordEncoder;
-  private final VerificationEmailCodeRepository verificationEmailCodeRepository;
+  private final VerificationEmailCodeStorage verificationEmailCodeStorage;
 
   @Transactional
   public void create(CreateUserCommand command) {
@@ -33,10 +37,23 @@ public class UserWriteService {
   }
 
   public void saveVerificationCode(SaveVerificationCodeCommand command) {
-    verificationEmailCodeRepository.save(command.email(), command.toEntity());
+    verificationEmailCodeStorage.save(command.email(), command.toEntity());
   }
 
   public void deleteVerificationCode(String email) {
-    verificationEmailCodeRepository.delete(email);
+    verificationEmailCodeStorage.delete(email);
+  }
+
+  public void verifyEmail(VerifyEmailCommand command) {
+    VerificationEmailCode code = verificationEmailCodeStorage.findByEmail(
+        command.email()).orElseThrow(
+        () -> UserException.from(INVALID_EMAIL_VERIFICATION));
+
+    if (!command.code().equals(code.getCode())) {
+      throw UserException.from(INVALID_EMAIL_VERIFICATION);
+    }
+
+    code.verified();
+    verificationEmailCodeStorage.verified(command.email(), code);
   }
 }
